Introduced atomic partition functions.
There may be a cleaner way of loading in the partition function tables, but it's not obvious to me and this is probably good enough for now.

I have locally verified that the code functions, but we should probably introduce some tests. For now I'm focused on trying to get this into the master branch so that we can work on this stuff in parallel.