Updated the documentation for hydrogenic_ff calculation and added test
I compared the equations for the hydrogenic ff opacity from the multiple reference sources (namely Kurucz 1970, Rybicki & Lightman 2004, Gray 2005) and was able to break the implementation proposed by Kurucz into a more standard format. In this commit I documented this relationship. The most important aspect of this is that I was able to document the part that corresponds to the free-free gaunt factor calculation (which we might replace in the future).

I also fixed errors in the names of the first arguments passed to ContinuumOpacity.H_I_ff and ContinuumOpacity.He_II_ff. The error related to the counter-intuitive naming convention of free-free transitions. Relatedly, I also added some additional documentation to ContinuumOpacity.hydrogenic_bf_opacity to try to clarify this confusion.

Additionally, I added a set of tests comparing our implementation of the free-free opacity problem to a secondary implementation taken from equation 8.10 from Gray (2005). This secondary formulation directly combines the opacity formula with the Saha equation and casts the constants into a completely different format.