Introduced tests for hydrogenic bf absorption
I added a set of tests comparing our implementation of the bound-free opacity to a secondary implementation taken from equation 8.8 of Gray (2005). The secondary formulation used different approximations for the cross-section and the contributions of bound-free transitions originating from higher energy levels. It also casts the constants into a completely different format.

Additionally, I added a set of tests to confirm that our integral approximation of the sum of the absorption contributions of transitions starting from elevated energy levels is consist with actually performing a direct summation. In the process, I slightly refactored the interfaces of ContinuumOpacity.H_I_bf and ContinuumOpacity.He_II_bf in a backwards compatible manner.