Fixed bug in Kurucz function for computing H2plus ff and bf opacity
I also introduced a test comparing the data against a table from Bates 1952 (which the polynomials listed in Kurucz may have been originally fit against). These polynomials agree with the tabulated data at a level of ~11.1%.

It's my intention to replace implementation of H2plus_bf_and_ff with the approach described in Gray (2005) in a subsequent commit since it has much better aggreement with the tabulated data. However, I wanted to have a working version of the Kurucz implementation to be part of the revision history.