Updated valid wavelength handling for ContinuumOpacity.Heminus_ff and added a test.
Added in a runtime check to make sure that valid wavelengths are passed into ContinuumOpacity.Heminus_ff (previously there was no check_. In the process, I realized that the documented range was slightly wrong so I fixed that too. Because the minimum valid wavelength is at 506.3 nm (in the middle of the visible wavelength regime), I commented the call to ContinuumOpacity.Heminus_ff in total_continuum_opacity

I also introduced a new test for ContinuumOpacity.Heminus_ff where we compare the opacities against the original tabulated values that polynomials in the function are supposed to approximate. While adding this test, I factored out some redunant code from this test, a test for SSSynth.ContinuumOpacity.Hminus_ff, and a test for ContinuumOpacity.H2plus_bf_and_ff