Altered ContinuumOpacity.Hminus_bf so that the cross-section is computed by directly interpolating data from Wishart (1979)
Previously, the cross-sections were computed using a polynomial from Gray (2005) that was fit to the tabulated data in Wishart (1979) while the tabulated data was used for verify the polynomial. Now, we use the Gray (2005) polynomial in our test to confirm that our transcription of the table is correct.

This change was primarily motivated by the fact directly using the table gives us opacities to wavelengths that are 130 nm longer.