Add experimental support for PHOENIX model atmospheres.
These have a couple differences from MARCS atmospheres that are relevant.
  - they lack a spatial coordinate
  - they have a reference wavelength that varies with effective temperature

The first difference is ok, because we can assume a plane-parallel goemetry and do radiative transfer with tau_scheme="anchored" (the default).

The second requires tracking the reference wavelength in the atmosphere object, and not assuming that it's 5000 \AA throughout the code.  Many of the changes in this commit are about keeping track of and accounting for this fact.

I have not included an internal linelist at 12,000 \AA, which is the value that PHOENIX uses besides 5000 \AA.  Instead, I've opted to require the user to supply a linelist that includes this region.